+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/feature/ToolStats.java
@@ -25,13 +25,13 @@ public class ToolStats extends Feature {
 
 	private static final List<String> toolsDurabilityDefault = Arrays.asList(
 			"minecraft:wooden_sword,1", "minecraft:wooden_pickaxe,1", "minecraft:wooden_shovel,1", "minecraft:wooden_hoe,1", "minecraft:wooden_axe,8",
-			"minecraft:stone_sword,1", "minecraft:stone_pickaxe,7", "minecraft:stone_shovel,48", "minecraft:stone_hoe,3", "minecraft:stone_axe,48",
+			"minecraft:stone_sword,1", "minecraft:stone_pickaxe,8", "minecraft:stone_shovel,48", "minecraft:stone_hoe,8", "minecraft:stone_axe,48",
 			"minecraft:iron_sword,375", "minecraft:iron_pickaxe,375", "minecraft:iron_shovel,375", "minecraft:iron_hoe,375", "minecraft:iron_axe,375",
 			"minecraft:elytra,144"
 	);
 
 	private static final List<String> toolEfficiencyDefault = Arrays.asList(
-			"minecraft:wooden_pickaxe,0.75", "minecraft:wooden_shovel,0.75", "minecraft:wooden_hoe,0.75", "minecraft:wooden_axe,0.75",
+			"minecraft:wooden_pickaxe,0.8", "minecraft:wooden_shovel,0.8", "minecraft:wooden_hoe,0.8", "minecraft:wooden_axe,0.8",
 			"minecraft:stone_pickaxe,0.85", "minecraft:stone_shovel,0.85", "minecraft:stone_hoe,0.85", "minecraft:stone_axe,0.85",
 			"minecraft:iron_pickaxe,0.9", "minecraft:iron_shovel,0.9", "minecraft:iron_hoe,0.9", "minecraft:iron_axe,0.9",
 			"minecraft:diamond_pickaxe,0.9", "minecraft:diamond_shovel,0.9", "minecraft:diamond_hoe,0.9", "minecraft:diamond_axe,0.9"